apps/api/pyproject.toml
[project]
name = "smallworld-api"
version = "0.1.0"
description = "Smallworld terrain analysis API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "numpy>=2.0",
    "scipy>=1.14",
    "pillow>=11.0",
    "pydantic-settings>=2.7",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/smallworld_api"]
